feat: extension-function to get number's ordinal added

+++ app/src/main/java/com/example/cryptoapp/data/model/cryptoCurrencyDetail/CoinDetailsResponse.kt
@@ -37,7 +37,7 @@ data class CoinDetailsResponse(
     @SerializedName("change")
     val change: String?,
     @SerializedName("rank")
-    val rank: Long?,
+    val rank: Int?,
     @SerializedName("sparkline")
     val sparkline: List<String>?,
     @SerializedName("allTimeHigh")
